step5 refactoring and add testcode
