+++ src/main/java/cargameracing/domain/Car.java
@@ -1,6 +1,4 @@
-package cargameracing;
-
-import java.util.Random;
+package cargameracing.domain;
 
 /**
  * Car Object
