+++ src/main/java/cargameracing/domain/MovingStrategy.java
@@ -1,4 +1,4 @@
-package cargameracing;
+package cargameracing.domain;
 
 @FunctionalInterface
 public interface MovingStrategy {
