+++ src/main/java/cargameracing/domain/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package cargameracing;
+package cargameracing.domain;
 
 import java.util.Random;
