+++ src/test/java/cargameracing/CarRacingGameTest.java
@@ -1,5 +1,6 @@
 package cargameracing;
 
+import cargameracing.domain.CarRacingGame;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,13 +14,19 @@ public class CarRacingGameTest {
 
     @BeforeEach
     public void setUp() {
-        racingGame = new CarRacingGame(Arrays.asList("tico", "bmw"), playCount);
+        racingGame = new CarRacingGame(Arrays.asList("tico", "bmw"), 10);
     }
 
     @Test
     @DisplayName("차의 개수 확인")
     void carCount() {
-        assertThat(racingGame.cars.length).isEqualTo(2);
+        assertThat(racingGame.getCars().length).isEqualTo(2);
         assertThat(racingGame.getCars()).hasSize(2);
+    }
+
+    @Test
+    @DisplayName("우승자 수 확인")
+    void winnerCount() {
+        assertThat(racingGame.filterWinners().size()).isGreaterThanOrEqualTo(1);
     }
 }
