+++ src/test/java/cargameracing/CarTest.java
@@ -1,5 +1,6 @@
 package cargameracing;
 
+import cargameracing.domain.Car;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
