+++ src/test/java/cargameracing/NameTest.java
@@ -1,5 +1,6 @@
 package cargameracing;
 
+import cargameracing.domain.Name;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
