+++ src/test/java/cargameracing/PositionTest.java
@@ -1,5 +1,6 @@
 package cargameracing;
 
+import cargameracing.domain.Position;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
